braswell: Remove feature and move DRM_I915_PRELIMINARY_HW_SUPPORT option
The only option in the "braswell" feature was
CONFIG_DRM_I915_PRELIMINARY_HW_SUPPORT=y. Preliminary hardware is a
moving target, not Braswell specific, so we move it to the i915 feature
which is a more appropriate place. This also gets us Skylake graphics
support in the common-pc BSPs.

Fixes [YOCTO 9319]

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>