Add edac support
Issue: LIN8-2810

1/6 [
Author: Tony Luck
Email: tony.luck@intel.com
Subject: sb_edac: support for Broadwell -EP and -EX
Date: Wed, 20 May 2015 19:10:35 -0300

commit fa2ce64f85be48003e8c88efcb8c3d7e4d0939ff upsteam

Basic support for the single socket Broadwell-DE processor
was added back in commit 1f39581a9a7a
   sb_edac: Add support for Broadwell-DE processor
This patch extends Broadwell support to cover the two
socket "-EP" and four socket "-EX" versions of Broadwell.
Only tested on the 2 socket - but this code is largely
cloned from the Haswell path.

Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>
Signed-off-by: chunguang yang <chunguang.yang@windriver.com>
]

2/6 [
Author: Tony Luck
Email: tony.luck@intel.com
Subject: sb_edac: Fix a typo and a thinko in address handling for Haswell
Date: Mon, 18 May 2015 17:39:06 -0300

commit bb89e7141a4e5690a046d3bdfa57b8521aab64b0 upstream

typo: "a7mode" chooses whether to use bits {8, 7, 9} or {8, 7, 6}
in the algorithm to spread access between memory resources. But
the non-a7mode path was incorrectly using GET_BITFIELD(addr, 7, 9)
and so picking bits {9, 8, 7}

thinko: BIT(1) of the dram_rule registers chooses whether to just
use the {8, 7, 6} (or {8, 7, 9}) bits mentioned above as they are,
or to XOR them with bits {18, 17, 16} but the code inverted the
test. We need the additional XOR when dram_rule{1} == 0.

Signed-off-by: Tony Luck <tony.luck@intel.com>
Acked-by: Aristeu Rozanski <aris@redhat.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>
Signed-off-by: chunguang yang <chunguang.yang@windriver.com>
]

3/6 [
Author: Tony Luck
Email: tony.luck@intel.com
Subject: sb_edac: Fix support for systems with two home agents per socket
Date: Mon, 18 May 2015 17:50:42 -0300

commit 7d375bffa52477c1be7db5a5dda99996582512c3 upstream

First noticed a problem on a 4 socket machine where EDAC only reported
half the DIMMS.  Tracked this down to the code that assumes that systems
with two home agents only have two memory channels on each agent. This
is true on 2 sockect ("-EP") machines. But four socket ("-EX") machines
have four memory channels on each home agent.

The old code would have had problems on two socket systems as it did
a shuffling trick to make the internals of the code think that the
channels from the first agent were '0' and '1', with the second agent
providing '2' and '3'. But the code didn't uniformly convert from
{ha,channel} tuples to this internal representation.

New code always considers up to eight channels.
On a machine with a single home agent these map easily to edac channels
0, 1, 2, 3. On machines with two home agents we map using:
  edac_channel = 4*ha# + channel
So on a -EP machine where each home agent supports only two channels
we'll fill in channels 0, 1, 4, 5, and on a -EX machine we use all of 0,
1, 2, 3, 4, 5, 6, 7.

[mchehab@osg.samsung.com: fold a fixup patch as per Tony's request and fixed
 a few CodingStyle issues]
Signed-off-by: Tony Luck <tony.luck@intel.com>
Acked-by: Aristeu Rozanski <aris@redhat.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>

Signed-off-by: chunguang yang <chunguang.yang@windriver.com>
]

4/6 [
Author: Tony Luck
Email: tony.luck@intel.com
Subject: EDAC: Remove arbitrary limit on number of channels
Date: Mon, 18 May 2015 17:32:37 -0300

commit c44696fff04ff62f65441afe9ea244b47653dd6d upstream

Currently set to "6", but the reset of the code will dynamically
allocate as needed.  We need to go to "8" today, but drop the check
completely to save doing this again when we need even larger numbers.

Signed-off-by: Tony Luck <tony.luck@intel.com>
Acked-by: Aristeu Rozanski <aris@redhat.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>
Signed-off-by: chunguang yang <chunguang.yang@windriver.com>
]

5/6 [
Author: Aristeu Rozanski
Email: aris@redhat.com
Subject: sb_edac: look harder for DDRIO on Haswell systems
Date: Fri, 12 Jun 2015 09:44:52 -0400

commit 7179385afef252cd3f52c0a06cc0c405ae6d66bc upstream

In case the memory banks are populated so the first channel isn't used, the
DDRIO PCI device won't be visible and it won't be possible to determine the
memory type.

Acked-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>
Signed-off-by: chunguang yang <chunguang.yang@windriver.com>
]

6/6 [
Author: Aristeu Rozanski
Email: aris@redhat.com
Subject: sb_edac: correctly fetch DIMM width on Ivy Bridge and Haswell
Date: Fri, 12 Jun 2015 15:08:17 -0400

commit 12f0721c5a70408e86257c5c99605cf743cd44c6 upstream

dimm_dev_type has been incorrectly determined in sb_edac. This patch fixes it
for Ivy Bridge and Haswell only since nothing like exists for Sandy Bridge.
We tested this patch in multiple systems matching the results with the
installed memory modules.

Acked-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>
Signed-off-by: chunguang yang <chunguang.yang@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>