ipv6: Extend the route lookups to low priority metrics
Issue: LIN6-10905

1/1 [
Author: Steffen Klassert
Email: steffen.klassert@secunet.com
Subject: ipv6: Extend the route lookups to low priority metrics.
Date: Tue, 28 Apr 2015 13:03:04 -0700

commit 9fbdcfaf97bf4b7d4fbd5b6a61f72167c73f37d9 upstream

We search only for routes with highest priority metric in
find_rr_leaf(). However if one of these routes is marked
as invalid, we may fail to find a route even if there is
a appropriate route with lower priority. Then we loose
connectivity until the garbage collector deletes the
invalid route. This typically happens if a host route
expires afer a pmtu event. Fix this by searching also
for routes with a lower priority metric.

Signed-off-by: Steffen Klassert <steffen.klassert@secunet.com>
Signed-off-by: Martin KaFai Lau <kafai@fb.com>
Reviewed-by: Hannes Frederic Sowa <hannes@stressinduktion.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Catalin Enache <catalin.enache@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>