bonding: don't use stale speed and duplex information
Issue: LIN6-10548

1/2 [
Author: Emil Tantilov
Email: emil.s.tantilov@intel.com
Subject: ixgbe: do not call check_link for ethtool in ixgbe_get_settings()
Date: Fri, 22 Jan 2016 10:56:09 +0800

In ixgbe_get_settings() the link status and speed of the interface
are determined based on a read from the LINKS register via the call
to mac.ops.check.link(). This can cause issues where external drivers
may end up with unknown speed when calling ethtool_get_setings().

Instead of calling the mac.ops.check_link() we can report the speed
from the adapter structure which is populated by the driver.

Signed-off-by: Emil Tantilov <emil.s.tantilov@intel.com>
Signed-off-by: yzhu1 <yanjun.zhu@windriver.com>
]

2/2 [
Author: Jay Vosburgh
Email: fubar@us.ibm.com
Subject: bonding: don't use stale speed and duplex information
Date: Mon, 29 Feb 2016 11:31:30 +0800

There is presently a race condition between the bonding periodic
link monitor and the updating of a slave's speed and duplex.  The former
occurs on a periodic basis, and the latter in response to a driver's
calling of netif_carrier_on.

    It is possible for the periodic monitor to run between the
driver call of netif_carrier_on and the receipt of the NETDEV_CHANGE
event that causes bonding to update the slave's speed and duplex.  This
manifests most notably as a report that a slave is up and "0 Mbps full
duplex" after enslavement, but in principle could report an incorrect
speed and duplex after any link up event if the device comes up with a
different speed or duplex.  This affects the 802.3ad aggregator
selection, as the speed and duplex are selection criteria.

    This is fixed by updating the speed and duplex in the periodic
monitor, prior to using that information.

    This was done historically in bonding, but the call to
bond_update_speed_duplex was removed in commit 876254ae2758 ("bonding:
don't call update_speed_duplex() under spinlocks"), as it might sleep
under lock.  Later, the locking was changed to only hold RTNL, and so
after commit 876254ae2758 ("bonding: don't call update_speed_duplex()
under spinlocks") this call is again safe.

Tested-by: "Tantilov, Emil S" <emil.s.tantilov@intel.com>
Cc: Veaceslav Falico <vfalico@gmail.com>
Cc: dingtianhong <dingtianhong@huawei.com>
Fixes: 876254ae2758 ("bonding: don't call update_speed_duplex() under spinlocks")
Signed-off-by: Jay Vosburgh <jay.vosburgh@canonical.com>
Acked-by: Ding Tianhong <dingtianhong@huawei.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
[read_lock/unlock does not make effect because of rtnl_lock/unlock.]
Signed-off-by: Zhu Yanjun <yzhu1@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>