intel-x86: drop HUGEPAGE settings for OVP.
It depends on !RT_FULL and for OVP8, we'll have that on most of the time.

Note that the rt.cfg of WR8 for the same BSP has these two items off.
So deleting them from the base BSP cfg isn't w/o precedent.

This will get rid of a kconfig audit warning during builds.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>