ipv6: Add missing route selection code
Issue: LIN5-21140

1/1 [
Author: Catalin Enache
Email: catalin.enache@windriver.com
Subject: ipv6: Add missing route selection code
Date: Fri, 18 Mar 2016 17:09:49 +0200

This is a follow-up patch to a4e902e ipv6: fix route selection
if kernel is not compiled with CONFIG_IPV6_ROUTER_PREF

We did a backport of an upstream commit but unfortunately this
wasn't complete, some upstream code is missing.

The initial test results were inaccurate and the initial
patch was integrated. It turns out that IPv6 route selection is not
working as expected after initial patch was ported.

With this patch we're completing the initial backporting.

Signed-off-by: Catalin Enache <catalin.enache@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>