bridge: Do not compile options in br_parse_ip_options
Issue: LIN7-5854

1/1 [
Author: Herbert Xu
Email: herbert@gondor.apana.org.au
Subject: bridge: Do not compile options in br_parse_ip_options
Date: Sat, 4 Oct 2014 22:18:02 +0800

commit 7677e86843e2136a9b05549a9ca47d4f744565b6 upstream

Commit 462fb2af9788a82a534f8184abfde31574e1cfa0

	bridge : Sanitize skb before it enters the IP stack

broke when IP options are actually used because it mangles the
skb as if it entered the IP stack which is wrong because the
bridge is supposed to operate below the IP stack.

Since nobody has actually requested for parsing of IP options
this patch fixes it by simply reverting to the previous approach
of ignoring all IP options, i.e., zeroing the IPCB.

If and when somebody who uses IP options and actually needs them
to be parsed by the bridge complains then we can revisit this.

Reported-by: David Newall <davidn@davidnewall.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Tested-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>
Signed-off-by: Li Zhou <li.zhou@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>