Revert "bsp/x86: import remaining sdhci-pci changes from ff tree."
This reverts commit fb53a8a979aa1be01d338d7e1f4dbfa4b49fa591.

These mmc changes are actually from yocto directly and are not BSP
specifc there, nor are they tied to any WR specific feature.

So rather than prop up the fraction that was here as was started
in the commit this reverts, we'll take the other approach and
get them out of this BSP specific location, and revisit making it
a global addition if there is WR demand for it.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>