powerpc64: kexec: fix the spin loop bug in kexec_wait
Issue: LIN8-3317

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: powerpc64: kexec: fix the spin loop bug in kexec_wait
Date: Thu, 12 May 2016 09:51:22 +0800

In the kexec_wait(), all the non-boot cpu spin on kexec_flag to wait
for the releasing of bootcpu. But we have used a wrong offset to load
the value of kexec_flag. Since the actual address we used to load the
value is in text code, the value is never null.  That means the non-boot
cpu never do the spin loop in the kexec_wait. Instead all the non-boot
cpu spin loop on the __secondary_hold_spinloop in __secondary_hold().
So in the following scenario, we would run into a TLB error for
instruction fetch on a SMT core.
       primary thread                  secondary thread
     kexec_smp_down()
       kexec_smp_wait()
         kexec_create_tlb()
                                       the 1:1 mapping tlb created by
                                       primary thread in kexec_create_tlb()
                                       may be overridden by this thread
         kexec_wait()
           bnea    0x60
             __secondary_hold()

So fix this by using the correct offset to load kexec_flag to make sure
that all the non-boot cpu are really waiting in kexec_wait() before the
boot cpu release them.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>