rt: drop rewrite of how two part irq handlers are dealt with
Issue: OVP8-36

The intlat benchmark, which for the context of this issue is essentially
a hardware timed response to IRQ events that are handled in a guest by
having the IRQ passed through from the host into the guest for the timer
device failed miserably upon moving to a 4.1-rt baseline.

The symptom was that once a busy loop task was put on the rt VCPU along
with the hardware IRQ of the timing device, the timing device would start
dropping IRQ events (normally 1kHz) and the reported response latencies
went through the roof.

Since the host kernel treats the rt VCPU as a single task, the
original guess was that the issue would be in the guest kernel, or
qemu, or in the intlat kernel module itself.  Logic being that it was
the guest kernel deciding if the IRQ got serviced ahead of the busy
loop loading task.

Oddly, bisecting (or swapping) components with a working OVP7/3.14
based system led to the host (not guest) kernel being the responsbile
culprit.

Complicating matters is that some earlier -rt kernels have a quirk
where assigning the affinity prior to having the irq actively taking
events doesn't take effect.  So unless you check /proc/interrupts of
the host and notice events in column zero vs. the CPU of the rt VCPU,
you can be misled into thinking things are working.

With all that in mind, working with a 3.18-rt baseline, it was found
that earlier versions were OK, but the most recent version was not.
Using the fast forward rt tags of linux-stable-rt, it was found that
the regression lied within this region:

9df6942317af Linux 3.18.27-rt26
5f0f37cdffa3 sched: Introduce the trace_sched_waking tracepoint
e068ec7031c1 irqwork: Move irq safe work to irq context
6983d62f5d22 net/core/cpuhotplug: Drain input_pkt_queue lockless
056c50bb9f02 net: Make synchronize_rcu_expedited() conditional on !RT_FULL
f339ae7c9600 dump stack: don't disable preemption during trace
171045bdda5e rtmutex: Use chainwalking control enum
83f58210e745 locking: locktorture: Do NOT include rwlock.h directly
7bb128748606 rtmutex: Handle non enqueued waiters gracefully
3ff367f166d4 ARM: smp: Move clear_tasks_mm_cpumask() call to __cpu_die()
374a09b1fcc9 genirq: Handle force threading of interrupts with primary and thread handler
642012ea1db5 cpufreq: Remove cpufreq_rwsem
c5734a9683d0 Linux 3.18.27-rt25

At this point, the trouble commit is fairly obvious, and it was
confirmed 374a was broken and 6420 was OK.

The original commit log was as follows:

    Author: Thomas Gleixner <tglx@linutronix.de>
    Date:   Sat Sep 19 11:56:20 2015 +0200

    genirq: Handle force threading of interrupts with primary and thread handler

    Force threading of interrupts does not deal with interrupts which are
    requested with a primary and a threaded handler. The current policy is
    to leave them alone and let the primary handler run in interrupt
    context, but we set the ONESHOT flag for those interrupts as well.

    Kohji Okuno debugged a problem with the SDHCI driver where the
    interrupt thread waits for a hardware interrupt to trigger, which cant
    work well because the hardware interrupt is masked due to the ONESHOT
    flag being set. He proposed to set the ONESHOT flag only if the
    interrupt does not provide a thread handler.

    Though that does not work either because these interrupts can be
    shared. So the other interrupt would rightfully get the ONESHOT flag
    set and therefor the same situation would happen again.

    To deal with this proper, we need to force thread the primary handler
    of such interrupts as well. That means that the primary interrupt
    handler is treated as any other primary interrupt handler which is not
    marked IRQF_NO_THREAD. The threaded handler becomes a separate thread
    so the SDHCI flow logic can be handled gracefully.

    The same issue was reported against 4.1-rt.

At this point, the host specific part of things start to make more
sense.  In the guest, the IRQ is just seen as a conventional IRQ, but
in the host, we have a more complicated situation, with pass through
IRQs showing up in the host as things like "kvm:0000:0a:00.1" - the
latter portion of the name being the PCI ID of the PCI device.

Looking at the commit, we see it shuffles a lot of specific affinity
related comments and code, and while at this point is isn't 100% clear
where it exactly goes wrong, one thing is very clear; the cure is in
this case worse than the disease.

While undeniably there may be issues with these slow SDHCI devices and
their split IRQ handlers, this "fix" causes regressions on existing
use cases and needs to be revisited.  We've had this SDHCI issue
present with nobody caring on on 3.2 (intel) 3.4 (OVP5), 3.6 (intel),
3.10 (OVP6) and 3.14 (OVP7).  So with that in mind, a revert/drop here
is not a significant concern.

The results found on the 3.18 baseline were confirmed to be forward
portable to the 4.1 baseline via a direct revert on the existing
tree.  Here in this change, we simply exclude the offending commit
from the rt series instead of reverting it.

Further investigation, specifically with explicit IRQF_NO_THREAD
deployment can be investigatged async, combined with looking for a
reproducer that isn't using the proprietary intlat code, so that the
upstream folks can be engaged.  But in the meantime, a drop/revert
is pretty much a no-brainer.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>