rt: relocate patches to proper directory
The ktypes dir is more for meta data content and not patches.
We already have features/rt for upstream rt content and then
features/wr-rt for content we add onto rt.  Stay with this
convention and move the patches in ktypes to wr-rt and mark
the file with a hint to avoid similar errors in the future.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>