hugetlb: delete dpdk including feature vs. cfg
Issue: OVP7-51

Newer kern tools trip over the duplicate feature entry (one here
and one in ktypes/standard) and try to apply the hugetlb patches
a second time (which of course fails).

For reference, the kern-tools commit that trips this is:

 From a03c100e813b1b5041bfc17c97f5bd4e72b795ad Mon Sep 17 00:00:00 2001
 From: Bruce Ashfield <bruce.ashfield@windriver.com>
 Date: Mon, 31 Aug 2015 15:55:33 -0400
 Subject: [PATCH] patching: only validate user supplied patches by default

We saw this earlier at treegen with newer tools, but now it breaks
builds due to an WR7 upgrade to the kernel tools.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>