ALSA: compress: fix an integer overflow check
Issue: SCP7-412

1/1 [
Author: Dan Carpenter
Email: dan.carpenter@oracle.com
Subject: ALSA: compress: fix an integer overflow check
Date: Wed, 16 Jul 2014 09:37:04 +0300

commit 6217e5ede23285ddfee10d2e4ba0cc2d4c046205 upstream

I previously added an integer overflow check here but looking at it now,
it's still buggy.

The bug happens in snd_compr_allocate_buffer().  We multiply
".fragments" and ".fragment_size" and that doesn't overflow but then we
save it in an unsigned int so it truncates the high bits away and we
allocate a smaller than expected size.

Fixes: b35cc8225845 ('ALSA: compress_core: integer overflow in snd_compr_allocate_buffer()')
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>