esdhc: Add new models to errata A-005055
Issue: LIN5-20652

1/3 [
Author: Haijun Zhang
Email: Haijun.Zhang@freescale.com
Subject: esdhc: Workaround for eSDHC clock glitch issue
Date: Mon, 14 Oct 2013 16:37:38 +0800

A-003980: SDHC: Glitch is generated on the card clock with software reset or
clock divider change
Description: A glitch may occur on the SDHC card clock when the software sets
the RSTA bit (software reset) in the system control register. It can also be
generated by setting the clock divider value. The glitch produced can cause
the external card to switch to an unknown state. The occurrence is not
deterministic.
Workaround:
A simple workaround is to disable the SD card clock before the software reset,
and enable it when the module resumes normal operation.
The Host and the SD card are in a master-slave relationship. The Host provides
clock and control transfer across the interface. Therefore, any existing
operation is discarded when the Host controller is reset.
The recommended flow is as follows:
1. Software disable bit[3], SDCLKEN, of the System Control Register
2. Trigger software reset and/or set clock divider
3. Check bit[3], SDSTB, of the Present State Register for stable clock
4. Enable bit[3], SDCLKEN, of the System Control Register
Using the above method, the eSDHC cannot send command or transfer data when

there is a glitch in the clock line, and the glitch does not cause any issue.
Signed-off-by: Haijun Zhang <haijun.zhang@freescale.com>
Change-Id: I1d8c3757ffdab73aff92cf1f0a6b89f82d11acfe
Reviewed-on: http://git.am.freescale.net:8181/4357
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Xie Xiaobo-R63061 <X.Xie@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>
[Original patch taken from QorIQ-SDK-V1.5-SOURCE-20131219-yocto.iso
 Change SDHCI_QUIRK2_DISABLE_CLOCK_BEFORE_RESET to (1<<8) to port 3.10
 kernel]
Signed-off-by: Bin Jiang <bin.jiang@windriver.com>

Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

2/3 [
Author: Haijun Zhang
Email: Haijun.Zhang@freescale.com
Subject: Powerpc:eSDCH: Add workaround about clock glitch issue on b4xxx
Date: Thu, 8 Aug 2013 09:22:25 +0800

A-003980: SDHC: Glitch is generated on the card clock with software reset or
clock divider change
Description: A glitch may occur on the SDHC card clock when the software sets
the RSTA bit (software reset) in the system control register. It can also be
generated by setting the clock divider value. The glitch produced can cause
the external card to switch to an unknown state. The occurrence is not
deterministic.

Workaround:
A simple workaround is to disable the SD card clock before the software reset,
and enable it when the module resumes normal operation.
The Host and the SD card are in a master-slave relationship. The Host provides
clock and control transfer across the interface. Therefore, any existing
operation is discarded when the Host controller is reset.
The recommended flow is as follows:
1. Software disable bit[3], SDCLKEN, of the System Control Register
2. Trigger software reset and/or set clock divider
3. Check bit[3], SDSTB, of the Present State Register for stable clock
4. Enable bit[3], SDCLKEN, of the System Control Register
Using the above method, the eSDHC cannot send command or transfer data when
there is a glitch in the clock line, and the glitch does not cause any issue.

Signed-off-by: Haijun Zhang <haijun.zhang@freescale.com>
Change-Id: I04ba0a6c107254b2f74f9c6ebd1f192d045448cf
Reviewed-on: http://git.am.freescale.net:8181/3858
Reviewed-by: Xie Xiaobo-R63061 <X.Xie@freescale.com>
Reviewed-by: Rivera Jose-B46482 <Jose.G.Rivera@freescale.com>
Tested-by: Rivera Jose-B46482 <Jose.G.Rivera@freescale.com>
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]
Signed-off-by: Bin Jiang <bin.jiang@windriver.com>
[Drop SVR_B4420 from quirk2 list since the current kernel does not support it]

Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

3/3 [
Author: Haijun Zhang
Email: Haijun.Zhang@freescale.com
Subject: esdhc: Add new models to errata A-005055
Date: Thu, 19 Dec 2013 11:21:33 +0800

A-005055: A glitch is generated on the card clock
due to software reset or a clock change.

Impact list:
T4240-4160-R1.0 B4860-4420-R1.0-R2.0 P3041-R1.0-R1.1-R2.0
P2041-2040-R1.0-R1.1-R2.0 P1010-1014-R1.0 P5020-5010-R1.0-R2.0
P5040-5021-R1.0-R2.0-R2.1

Signed-off-by: Haijun Zhang <haijun.zhang@freescale.com>
Change-Id: Idf4e1fd52ec4b5748cf9cbf86365a70c9952d1be
Reviewed-on: http://git.am.freescale.net:8181/5925
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Xiaobo Xie <X.Xie@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]
Signed-off-by: Bin Jiang <bin.jiang@windriver.com>
[Drop unsupported SOC Revisions from quirk2 list]

Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>