ALSA: usb-audio: Fix NULL dereference in create_fixed_stream_quirk()
Issue: LIN5-21502

1/3 [
Author: Mark Hills
Email: mark@xwax.org
Subject: ALSA: usb-audio: Trust fields given in the quirk
Date: Sun, 17 Mar 2013 11:07:54 +0000

commit 59ea586f54f27fde1202214f1525713356a44918 upstream

The maxpacksize field is given in some quirks, but it gets ignored (in
favour of wMaxPacketSize from the first endpoint.) This patch favours
the one in the quirk.

Digidesign Mbox and Mbox 2 are the only affected quirks and the devices
are assumed to be working without this patch. So for safety against the
values in the quirk being incorrect, remove them.

The datainterval is also ignored but there are not currently any quirks
which choose to override this.

Cc: Damien Zammit <damien@zamaudio.com>
Cc: Chris J Arges <christopherarges@gmail.com>
Signed-off-by: Mark Hills <mark@xwax.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

2/3 [
Author: Eldad Zack
Email: eldad@fogrefinery.com
Subject: ALSA: usb-audio: fix regression for fixed stream quirk
Date: Tue, 9 Jul 2013 20:36:15 +0200

commit 42d4ab83 upstream

Commit 8f898e92aea2c24c7f379ee265d178f69ebb9c07 removed the redundant
reads of bInterfaceProtocol from the descriptors, but introduced a
regression to devices with quirks of type QUIRK_AUDIO_FIXED_ENDPOINT,
since fp->protocol is not set in setup process.

As a consequence, audio streams would not get initialized, as the
following logs show:

[   48.923043] setting usb interface 3:1
[   48.923056] Creating new capture data endpoint #81
[   48.923484] 4:3:1: cannot set freq 48000 to ep 0x81

This patch sets fp->protocol in create_fixed_stream_quirk() and
resolves the regression.

Signed-off-by: Eldad Zack <eldad@fogrefinery.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

3/3 [
Author: Takashi Iwai
Email: tiwai@suse.de
Subject: ALSA: usb-audio: Fix NULL dereference in create_fixed_stream_quirk()
Date: Tue, 15 Mar 2016 12:09:10 +0100

commit 0f886ca12765d20124bd06291c82951fd49a33be upstream

create_fixed_stream_quirk() may cause a NULL-pointer dereference by
accessing the non-existing endpoint when a USB device with a malformed
USB descriptor is used.

This patch avoids it simply by adding a sanity check of bNumEndpoints
before the accesses.

Bugzilla: https://bugzilla.suse.com/show_bug.cgi?id=971125
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>