powerpc/tm: Block signal return setting invalid MSR state
Issue: SCP6-762

1/1 [
Author: Michael Neuling
Email: mikey@neuling.org
Subject: powerpc/tm: Block signal return setting invalid MSR state
Date: Thu, 19 Nov 2015 15:44:44 +1100

commit d2b9d2a5ad5ef04ff978c9923d19730cb05efd55 upstream

Currently we allow both the MSR T and S bits to be set by userspace on
a signal return.  Unfortunately this is a reserved configuration and
will cause a TM Bad Thing exception if attempted (via rfid).

This patch checks for this case in both the 32 and 64 bit signals
code.  If both T and S are set, we mark the context as invalid.

Found using a syscall fuzzer.

Fixes: 2b0a576d15e0 ("powerpc: Add new transactional memory state to the signal context")
Cc: stable@vger.kernel.org # v3.9+
Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Yue Tao <yue.tao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>