spi: zynq-qspi: Wait for FIFO to be empty when switching TXD registers
Issue: LIN7-6375

1/1 [
Author: Harini Katakam
Email: harinik@xilinx.com
Subject: spi: zynq-qspi: Wait for FIFO to be empty when switching TXD registers
Date: Mon, 25 Jul 2016 17:49:10 +0800

This patch comes from:
      https://github.com/Xilinx/linux-xlnx.git

When switching from TXD0 to TXD1/2/3, software has to wait for FIFO to be
empty before writing to TXD1/2/3 as per the documentation.

Although previous discussions and tests revealed that this waiting
is required only for TXD1/2/3 to TXD0 switches and not vice-verse,
design team conveyed that TXD0 to TXD1/2/3 switching can behave unpredictably
if wait is not used. Hence this change is made to comply.

Signed-off-by: Harini Katakam <harinik@xilinx.com>
Acked-by: Punnaiah Choudary Kalluri <punnaia@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Zhixiong Chi <Zhixiong.Chi@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>