ALSA: hrtimer: Fix stall by hrtimer_cancel()
Issue: LIN6-11179

1/1 [
Author: Takashi Iwai
Email: tiwai@suse.de
Subject: ALSA: hrtimer: Fix stall by hrtimer_cancel()
Date: Mon, 18 Jan 2016 13:52:47 +0100

commit 2ba1fe7a06d3624f9a7586d672b55f08f7c670f3 upstream

hrtimer_cancel() waits for the completion from the callback, thus it
must not be called inside the callback itself.  This was already a
problem in the past with ALSA hrtimer driver, and the early commit
[fcfdebe70759: ALSA: hrtimer - Fix lock-up] tried to address it.

However, the previous fix is still insufficient: it may still cause a
lockup when the ALSA timer instance reprograms itself in its callback.
Then it invokes the start function even in snd_timer_interrupt() that
is called in hrtimer callback itself, results in a CPU stall.  This is
no hypothetical problem but actually triggered by syzkaller fuzzer.

This patch tries to fix the issue again.  Now we call
hrtimer_try_to_cancel() at both start and stop functions so that it
won't fall into a deadlock, yet giving some chance to cancel the queue
if the functions have been called outside the callback.  The proper
hrtimer_cancel() is called in anyway at closing, so this should be
enough.

Reported-and-tested-by: Dmitry Vyukov <dvyukov@google.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>